requirements.txt
asgiref==3.4.1
astroid==2.8.5
autopep8==1.6.0
black==21.10b0
click==8.0.3
Django==3.2.7
djangorestframework==3.12.4
flake8==4.0.1
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2021.1
regex==2021.11.10
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.2
typing-extensions==3.10.0.2
wrapt==1.13.3
